extract wrapper into folder
