+++ libs/service/service.go
@@ -1,4 +1,4 @@
-package utils
+package service
 
 import "sync/atomic"
 
@@ -9,21 +9,17 @@ const (
 	Running
 )
 
-// StateWrapper provides an interface to change an check service state
-type StateWrapper struct {
+// Service provides an interface to change an check service state
+type Service struct {
 	state uint32
 }
 
-func NewStateWrapper() *StateWrapper {
-	return &StateWrapper{}
-}
-
 // SetState allows to change service state
-func (s *StateWrapper) SetState(state State) {
+func (s *Service) SetState(state State) {
 	atomic.StoreUint32(&s.state, uint32(state))
 }
 
 // State returns service current state
-func (s *StateWrapper) State() State {
+func (s *Service) State() State {
 	return State(atomic.LoadUint32(&s.state))
 }
